+++ packages/evokit-button/CHANGELOG.md
@@ -1,5 +1,9 @@
 [README]: /packages/evokit-button/README.md
 
+[button-size]: /packages/evokit-button/?id=button-size
+[button-theme]: /packages/evokit-button/?id=button-theme
+[button-valign]: /packages/evokit-button/?id=button-valign
+[button-events]: /packages/evokit-button/?id=button-events
 
 # EvoKit - Button
 
@@ -10,6 +14,15 @@
 
 ## CHANGELOG
 
-### 3.0.0-beta.0 *(21.02.2020)*
+### 3.0.0-beta.1 *(02.03.2020)*
+
+- **Added** new prop [`button-size`][button-size]
+- **Added** new prop [`button-valign`][button-valign]
+- **Added** new prop [`button-events`][button-events]
+- **Added** additional states for theme [`button-theme`][button-theme]
+- **Fixed** tap highlight color if [`button-theme`][button-theme] is not set
+- **Fixed** max width
+
+### 3.0.0-beta.0 *(20.02.2020)*
 
 - Initial version.